+++ hexfont.h
@@ -34,7 +34,7 @@ extern "C" {
 
 // An individual character in the font
 typedef struct hexfont_character {
-    uint16_t codepoint;
+    uint32_t codepoint;
     uint8_t *glyph;
     size_t glyph_len;
     uint8_t width;
@@ -60,7 +60,7 @@ typedef struct hexfont {
 
 hexfont * const hexfont_load(const char *file, const uint8_t glyph_height);
 hexfont * const hexfont_load_data(const char *data, const uint8_t glyph_height);
-hexfont_character * const hexfont_get(hexfont * const font, const uint16_t codepoint);
+hexfont_character * const hexfont_get(hexfont * const font, const uint32_t codepoint);
 const bool hexfont_character_get_pixel(hexfont_character * const c, const size_t x, const size_t y);
 void hexfont_destroy(hexfont * const font);
 void hexfont_dump_character(hexfont_character * const c, FILE *fp);
